Added authorization and security around the Track and session controller.
--HG--
extra : convert_revision : svn%3A25cf9c4d-fa41-0410-a78d-8fe958b11fb0/trunk%40798

+++ src/UI/Controllers/TrackController.cs
@@ -2,6 +2,7 @@ using System;
 using System.Web.Mvc;
 using CodeCampServer.Core.Domain;
 using CodeCampServer.Core.Domain.Model;
+using CodeCampServer.Core.Services;
 using CodeCampServer.UI.Helpers.Filters;
 using CodeCampServer.UI.Helpers.Mappers;
 using CodeCampServer.UI.Models.Forms;
@@ -12,16 +13,17 @@ namespace CodeCampServer.UI.Controllers
 	public class TrackController : SaveController<Track, TrackForm>
 	{
 		private readonly ITrackRepository _repository;
-
 		private readonly ITrackMapper _mapper;
 		private readonly ISessionRepository _sessionsRepository;
+	    private readonly ISecurityContext _securityContext;
 
-		public TrackController(ITrackRepository repository, ITrackMapper mapper, ISessionRepository sessionsRepository)
+	    public TrackController(ITrackRepository repository, ITrackMapper mapper, ISessionRepository sessionsRepository, ISecurityContext securityContext)
 			: base(repository, mapper)
 		{
 			_repository = repository;
 			_mapper = mapper;
 			_sessionsRepository = sessionsRepository;
+		    _securityContext = securityContext;
 		}
 
 		public ViewResult New(Conference conference)
@@ -38,12 +40,21 @@ namespace CodeCampServer.UI.Controllers
 
 		public ViewResult Edit(Track track)
 		{
+            if(!_securityContext.HasPermissionsFor(track.Conference))
+            {
+                return NotAuthorizedView;
+            }
 			return View(_mapper.Map(track));
 		}
 
 		[ValidateModel(typeof (TrackForm))]
 		public ActionResult Save([Bind(Prefix = "")] TrackForm trackForm, Conference conference, string urlreferrer)
 		{
+            if(!_securityContext.HasPermissionsFor(conference))
+            {
+                return NotAuthorizedView;
+            }
+
 			Func<Track, ActionResult> successRedirect = GetSuccessRedirect(conference, urlreferrer);
 
 			return ProcessSave(trackForm, successRedirect);
@@ -61,8 +72,13 @@ namespace CodeCampServer.UI.Controllers
 			return successRedirect;
 		}
 
-		public RedirectToRouteResult Delete(Track track)
+		public ActionResult Delete(Track track)
 		{
+            if(!_securityContext.HasPermissionsFor(track.Conference))
+            {
+                return NotAuthorizedView;
+            }
+
 			if (_sessionsRepository.GetAllForTrack(track).Length == 0)
 			{
 				_repository.Delete(track);
